refactor(DTO):substitui entidade diretamente exposta por uma DTO

+++ src/main/java/com/estagioprojeto/service/impl/TutorialServiceImpl.java
@@ -1,12 +1,14 @@
 package com.estagioprojeto.service.impl;
 
+import com.estagioprojeto.dto.TutorialDto;
+import com.estagioprojeto.mapper.TutorialMapper;
 import com.estagioprojeto.model.Tutorial;
 import com.estagioprojeto.repository.TutorialRepository;
 import com.estagioprojeto.service.TutorialService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestParam;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,14 +16,16 @@ import java.util.Optional;
 public class TutorialServiceImpl implements TutorialService {
 
     private final TutorialRepository tutorialRepository;
+    private final TutorialMapper tutorialMapper;
 
-    public TutorialServiceImpl(TutorialRepository tutorialRepository) {
+    public TutorialServiceImpl(TutorialRepository tutorialRepository, TutorialMapper tutorialMapper) {
         this.tutorialRepository = tutorialRepository;
+        this.tutorialMapper = tutorialMapper;
     }
-
+    //DTO
     @Override
-    public List<Tutorial> findPorPublicacao(boolean opcao) {
-        List<Tutorial> respTutorials = tutorialRepository.findByPublished(opcao);
+    public List<TutorialDto> findPorPublicacao(boolean opcao) {
+        List<TutorialDto> respTutorials = tutorialMapper.toDtoList(tutorialRepository.findByPublished(opcao));
 
         if (respTutorials.isEmpty()) {
             return respTutorials;
@@ -31,64 +35,70 @@ public class TutorialServiceImpl implements TutorialService {
 
     @Override
     public void deletarPorId(long id) {
+        if (!tutorialRepository.existsById(id)) {
+            throw new EntityNotFoundException("Tutorial não encontrado com o ID: " + id);
+        }
         tutorialRepository.deleteById(id);
     }
 
+    //DTO
     @Override
-    public List<Tutorial> chamarTodosPorDescricao(String description) {
-        return tutorialRepository.findAllByDescription(description);
+    public List<TutorialDto> chamarTodosPorDescricao(String description) {
+        return tutorialMapper.toDtoList(tutorialRepository.findAllByDescription(description));
     }
-
+    //DTO
     @Override
-    public Tutorial atualizarPorId(Tutorial tutorial, long id){
-        Optional<Tutorial> tutorialData = tutorialRepository.findById(id);
-
-        if (tutorialData.isPresent()) {
-            Tutorial _tutorial = tutorialData.get();
-            _tutorial.setTitle(tutorial.getTitle());
-            _tutorial.setDescription(tutorial.getDescription());
-            _tutorial.setPublished(tutorial.isPublished());
-            return tutorialRepository.save(_tutorial);
+    public TutorialDto atualizarPorId(long id, TutorialDto tutorialDto) {
+        try {
+            if(tutorialRepository.existsById(id)){
+                tutorialMapper.updateEntityFromDto(tutorialRepository.findById(id).get(), tutorialDto);
+                return tutorialMapper.toDto(tutorialRepository.save(tutorialRepository.findById(id).get()));
+            }
+        }catch (Exception e){
+            e.printStackTrace();
         }
         return null;
     }
-
+    //DTO
     @Override
-    public Tutorial criarTutorial(Tutorial tutorial){
-        tutorialRepository.save(new Tutorial(tutorial.getTitle(), tutorial.getDescription(), true));
+    public TutorialDto criarTutorial(TutorialDto tutorial){
+        Tutorial tutorialEntity = tutorialMapper.toEntity(tutorial);
+        Tutorial tutorialSalvo = tutorialRepository.save(new Tutorial(tutorialEntity.getTitle(),
+                tutorialEntity.getDescription(),
+                true));
 
         try{
-            if(tutorial!=null){
-                return tutorial;
-            }
+            return tutorialMapper.toDto(tutorialSalvo);
 
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
-        return null;
 
     }
-
+    //DTO
     @Override
-    public Optional<Tutorial> localizarPorId(long id){
-        if(tutorialRepository.findById(id).isPresent()){
-            return  tutorialRepository.findById(id);
+    public TutorialDto localizarPorId(long id){
+        try {
+            if (tutorialRepository.findById(id).isPresent()) {
+                TutorialDto tutorialLocalizado = tutorialMapper.toDto(tutorialRepository.findById(id).get());
+                return tutorialLocalizado;
+            }
+        }catch (Exception e){
+            throw new RuntimeException(e);
         }
-        return Optional.empty();
+        return null;
     }
-
+    //DTO
     @Override
-    public List<Tutorial> localizarTodos(String title){
-        List<Tutorial> tutorials = new ArrayList<Tutorial>();
+    public List<TutorialDto> localizarTodos(String title){
+        //List<Tutorial> tutorials = new ArrayList<Tutorial>();
 
         if (title == null) {
-            tutorials.addAll(tutorialRepository.findAll());
+            return tutorialMapper.toDtoList(tutorialRepository.findAll());
         }
         else {
-            tutorials.addAll(tutorialRepository.findByTitleContaining(title));
+            return tutorialMapper.toDtoList(tutorialRepository.findByTitleContaining(title));
         }
-
-        return tutorials;
     }
